requirements.txt
git+https://github.com/Rapptz/discord.py
python-dotenv
async-timeout==3.0.1
requests
urllib3